Added loading skeleton to AdsCard component
Enhancements have been made to the handling of 'loading' states in the community advertising space, specifically in the 'AdsCard' component. Added the 'AdsCardSkeleton' component to provide a more engaging and informative experience for end users during periods of data fetching or latency. This loading skeleton appears in place of active ads until the data becomes available.

This change utilizes the 'isLoading' state from the 'CommunityContext'. In effect, as the data loads, we present users with a skeleton UI, contributing to an overall smoother, more seamless experience.

No referenced issues.