Adjusted element z-index, Populated image URLs,
add in server imageURl data

In this update, we've made a variety of changes to enhance the performance and usability of the application.

First off, we've adjusted the 'z-index' property of certain elements in the profile CSS file. This adjustment helps to fine-tune the stacking order of the elements, thus delivering a better user interface.

We've also modified the MapEditor file to appropriately populate image URLs for various users, replacing the global 'profileData' object with specific user objects. This update ensures that the correct images are displayed for each user.