Add responsive styling to CreateAd page; implement dynamic profile icon
An extensive overhaul was conducted on the CSS for the CreateAd component, implementing responsiveness for better display on various screen sizes. Styling and layout adjustments improve usability and appearance on medium-sized devices, enhancing the user experience.

In the header components, a redirection link to the 'create ad' page was attached to the 'plus' icon, enhancing navigational ease. Padding tweaks refined the display of said icon.

In the main Header, a function was introduced to dynamically select profile icons based on user-defined gender, fostering inclusivity and personalization.

Lastly, a console log was added to the UserContext for development purposes, aiding debugging and state tracking.

Keep in mind to remove the debugging statements before final deployment to retain good code etiquette and maintain a clean production environment.