Updated meta tags and title in index.html for SEO
The latest changes primarily focus on improvising SEO for the Pensa Club platform. The HEAD tag in the index.html was updated with a more descriptive title and meta description that better summarize the purpose and functionality of the platform. Inclusion of relevant keywords supports search engine optimization. This will improve the discoverability of the platform and its mission of aiding elderly people in Bulgaria.