Refactor ad creation, improve location fetching logic
 Reorganized the useEffect dependencies to streamline the fetching of town data. Rather than depending on the entire settlements list to update, the new change isolates town fetching to occur only when the selected region and subregion states change.

Updated region, subregion, and town selection setters to retain their current state if a new selection does not exist. This prevents overwriting with undefined selections.

Finally, redundant white spaces were removed from the ad creation view, and an unnecessary console error in the CommunityContext was commented out.

Now fetchTowns only runs if and when there are changes to selectedRegion and selectedSubregion, making the operation smoother and more efficient. This adjustment makes the code more efficient and easier to maintain for further updates.