Expanded multi-language support and improved UX in map view
This comprehensive update includes a variety of changes to better support users. User experience in map view is significantly improved. In addition to this, several new pieces of text have been added to the English and Bulgarian locale files, enhancing the multi-language functionality of the application. The map editor CSS was also slightly changed, adding new styling options and improving the current ones. A minor bug was caught and fixed in the adsController.js file. An unnecessary status update attempted through the ad-create endpoint has been taken care of to conform to best practices. We will continue to make necessary updates to enhance user satisfaction. This update directly improves the user's ability to navigate and interact in the map view.