Enhance user account and ad management features
This commit significantly enhances the management of user accounts and ads. To start with, it includes ads' data of a user into the user's details which will give a more complete profile of a user. In addition, it also improves the verification method for ad status update requests, ensuring that an admin comment must accompany denied ads. Furthermore, better handling is provided for ad status update cases if the ad is already in the requested status. In the case of editing an ad, restrictions are set up to prevent status changes through the endpoint.