Expanded admin dashboard with multi-language support and improved functionality
This update focuses on enhancing the administrative functionality of the platform and adding multi-language translation support to several areas.

Firstly, the Admin Dashboard is further enhanced with more control options. The changes allow admins to Approve, Reject, or Delete a pending announcement with accurate translations provided. Also, single and multiple count conditions for ads are addressed to fix a grammatical inconsistency.

Then, the codebase has translations added to the Flyout and PendingAnnouncements components to ensure that language settings are honored uniformly across the dashboard.

Notification messages are also implemented to signal success when an ad is approved or rejected, ensuring better feedback to the user about their actions' results.

Lastly, formatting and code style improvements are made across several files to enhance code readability and maintainability.

These optimizations make the administration of the site more comfortable and more efficient, especially for non-English speaking admins.

The changes don't affect anything functionally on the front-end for regular users. However, it aligns with our dedication to ensuring a more universal and inclusive experience across multiple regions by enabling vital translations.