Add paginated and endless scrolling ad results on the community page
Instead of loading all the advertisements at once, we've incorporated endless scrolling and pagination to the community page. This will ensure a smoother experience for the user and reduce load time.

To achieve this, an adjustment was made to both the frontend to handle the scrolling events and the backend to paginate the data being sent from the server. The lodash.debounce package was also added to handle multiple triggering of load events when user is scrolling.

In addition, the server-side code has been updated to return an indication of whether more ads are to be fetched for the current search options.

Other refactoring and improvements include changes related to handling search-ads and loading ads. Also added, a resolve functionality to suggested user to handle resolving of user suggestion. Finally, fixed a minor bug which showed the loading skeletons even when there are no ads to fetch. Now the skeletons will only appear when there are no ads and the system is fetching new ones.

These changes will improve performance, especially for cases with a large number of results.