Added comment field for admin role changes and refined data caching
The update pertains to the implementation of a new feature and enhancement of the data caching mechanism. A comment field was added to keep track of admin-imposed user role changes for accountability and context. Admins can now add a comment detailing why a user's role is being modified. ValidationError mechanics ensure that the comment must be 10-1000 characters long. Alterations were made to the user_account model, admin controller, and a migration file was added.

The caching mechanism was improved for more precise data storage. Specifically, user email was added as an additional parameter in memoryCache for facilitating efficient retrieval of user-specific ads data.

Also minor tweaks regarding data retrieval and editing in the ads controller were executed for smoother user experience. Email has been omitted in ad details and safeguards were added to prevent unauthorized modifications.