Add comment feature in user suggestion and internationalize the feature details
Extended the functionality of user suggestions with a new comment feature. These are now formatted and sorted by date to improve UX in the admin dashboard. Added corresponding comment functionalities in Bulgarian and English translation. Ensured mobile responsiveness was maintained throughout the changes. Modified how text is truncated in the admin dashboard to improve data readability and adjusted CSS for comment modals and tables for better visuals. Minor adjustments were also made to file imports and readability of certain functions.

This enhancement eases admin's ability to manage and respond to user suggestions, fostering better engagement with the user community.