Refine password regex in signup utility
Simplified password regular expression (regex) in the signup utility. Previously, the regex necessitated a mixed case, but now it merely calls for at least one alphabet character, irrespective of the case. This change makes the password requirement slightly more lenient and user-friendly, while maintaining a level of complexity with the demand for numeric characters. It can be part of a broader strategy to increase user signup by introducing less restrictive rules.