Implemented improved user navigation in MapEditor
In response to user feedback on the MapEditor interface, a few changes were made to improve user experience and navigation. Primarily, the changes focus on bringing more fluid and intuitive interacting behavior to the MapEditor by smoothly scrolling into view and adjusting viewport positions.

One key new feature is that MapEditor now automatically scrolls to a relevant view on selecting read-more information of a marker on the map. This aids users to focus on the content they chose to see without manual viewport adjustment. If the sidebar opens due to a marker's read-more option being selected, the page will auto-scroll to brought the map view to the start. Meanwhile, clicking outside the sidebar closes it and returns the viewport to the previous scroll position, maintaining the users' perspective on the map. These changes overall better the user navigation experience and assist in retaining the map context while interacting with the elements.

Kindly note that the functionality to smoothly scroll the sidebar into view on selecting a user has been commented out at the moment. Unforeseen scroll behavior was observed which will need debugging in the future.

Resolves: #452

Co-authored-by: Borislav Iliev <85155699+bebo13133@users.noreply.github.com>