Added scroll functionality to AdDetails view
This update adds code for managing the scroll position in the AdDetail's component. The useRef hook is now imported and a reference to the adContainer is set. This reference is used to automatically smoothly scroll to that position when the community sidebar is populated with a user's ads. Additionally, a new state variable is established to remember the original scroll position, allowing users to return to their initial page position when they close the sidebar. This improves the UX by ensuring that users don't lose their place on the page when interacting with the UI.

No related issues are referenced in this commit.

Co-authored-by: Borislav Iliev <85155699+bebo13133@users.noreply.github.com>
Co-authored-by: Dafy Yasenova <dafy_yasenova@abv.bg>