Improved validation and i18n in admin comments
This commit introduces error handling on comment submission in the admin dashboard. Now, admins won't be able to submit blank comments - an error message is displayed if attempted. This avoids potential confusion and assists admins in managing user suggestions effectively.

The language of search prompts has also been enhanced for a better user experience, especially for Bulgarian and English users. Instead of asking for a "tag", it now asks for a "keyword", making it easier for users to understand what they need to search.

References: No related issues.

Co-authored-by: Borislav Iliev <85155699+bebo13133@users.noreply.github.com>
Co-authored-by: Dafy Yasenova <dafy_yasenova@abv.bg>