Update various input field validations and update translations
This commit enhances input field validations in various forms and components. Regex patterns were updated to strengthen criteria for names and a new streetRegex was added for validating street inputs. The new regex patterns should ensure more robust, consistent, and accurate data from users.

Translations are updated to reflect additional invalid input response messages in both Bulgarian and English versions. The message displays when the user inputs an invalid street in the profile form.

Revisions were made in the component styles to improve visual response on error in forms.

A conditional update is also made in the telephone number validation to prevent it from validation-required error when it is not a required field in the form.

Additionally, this commit migrates the codebase to react-i18next v14.1.3 from v14.1.2, to benefit from the bug fixes, optimizations, and new features introduced in this minor version. The package-lock.json and package.json files have been updated in both areas to reflect this change.

Co-authored-by: Borislav Iliev <85155699+bebo13133@users.noreply.github.com>
Co-authored-by: Dafy Yasenova <dafy_yasenova@abv.bg>